requirements.txt
absl-py==1.4.0
accelerate==0.31.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
altair==5.3.0
aniso8601==9.0.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==22.2.0
aura-sr==0.0.2
av==12.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bitsandbytes==0.43.3
bleach==6.0.0
blinker==1.8.2
boto3==1.35.0
botocore==1.35.0
brotlipy==0.7.0
cachetools==5.3.0
chumpy==0.70
click==8.1.7
cloudpickle==3.0.0
coloredlogs==15.0.1
comm==0.1.2
contourpy==1.0.7
controlnet-aux==0.0.9
cycler==0.11.0
databricks-sdk==0.30.0
debugpy==1.6.6
decorator==4.4.2
decord==0.6.0
defusedxml==0.7.1
Deprecated==1.2.14
diffusers==0.29.2
dnspython==2.6.1
docker==7.1.0
docopt==0.6.2
einops==0.8.0
email_validator==2.2.0
entrypoints==0.4
executing==1.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.16.2
ffmpy==0.3.2
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.16.0
google-auth-oauthlib==0.4.6
gradio==4.37.2
gradio_client==1.0.2
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.51.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib==1.0.4
importlib_metadata==7.2.1
importlib_resources==6.4.0
ipykernel==6.20.2
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.31
jaxlib==0.4.31
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.4.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.0.1
jupyter_core==5.1.5
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab==3.5.3
jupyterlab-language-pack-zh-CN==3.5.post4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.19.0
kiwisolver==1.4.4
kornia==0.7.3
kornia_rs==0.1.4
lazy_loader==0.4
lightning-utilities==0.11.6
Mako==1.3.5
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
matrix-client==0.4.0
mdurl==0.1.2
mediapipe==0.10.14
mistune==2.0.4
ml-dtypes==0.4.0
mlflow==2.15.1
mlflow-skinny==2.15.1
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.16.4
nbformat==5.7.3
nest-asyncio==1.5.6
netron==7.8.4
networkx==3.3
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.16.2
onnxruntime==1.18.1
onnxruntime-gpu==1.18.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opt-einsum==3.3.0
orjson==3.10.5
packaging==23.0
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.5.0
platformdirs==2.6.2
proglog==0.1.10
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==5.28.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==2.8.0
pydantic_core==2.20.0
pyDeprecate==0.3.2
pydub==0.25.1
PyGithub==2.3.0
Pygments==2.14.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.4
python-multipart==0.0.9
pytorch-lightning==1.6.5
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
querystring-parser==1.2.4
regex==2024.5.15
requests==2.32.3
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rsa==4.9
Rtree==1.3.0
ruff==0.5.0
s3transfer==0.10.2
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
semantic-version==2.10.0
Send2Trash==1.8.0
shellingham==1.5.4
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.5.0
soundfile==0.12.1
soupsieve==2.3.2.post1
spandrel==0.3.4
SQLAlchemy==2.0.32
sqlparse==0.5.1
stack-data==0.6.2
starlette==0.37.2
sympy==1.12.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
threadpoolctl==3.5.0
tifffile==2024.8.10
timm==0.6.7
tinycss2==1.2.1
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
torchmetrics==1.4.1
torchsde==0.2.6
tornado==6.2
traitlets==5.8.1
trampoline==0.1.2
transformers==4.42.3
trimesh==4.4.6
triton==3.0.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uri-template==1.2.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.0
websockets==11.0.3
Werkzeug==3.0.3
widgetsnbextension==4.0.5
wrapt==1.16.0
xformers==0.0.27.post2
yacs==0.1.8
yarg==0.1.9
yarl==1.9.4
zipp==3.19.2